x86: Keep fp_regs.h standalone
We cannot include stdint.h in fp_regs.h since this is included by
mach_types.h for bootstrapping, before we have stdint.h from glibc.

* i386/include/mach/i386/fp_reg.h: Do not include <stdint.h>
(XSAVE_XCOMP_BV_COMPACT, struct i386_xfp_xstate_header): Use unsigned
long long instead of uint64_t.

+++ i386/include/mach/i386/fp_reg.h
@@ -27,8 +27,6 @@
 #ifndef	_MACH_I386_FP_REG_H_
 #define	_MACH_I386_FP_REG_H_
 
-#include <stdint.h>
-
 /*
  *	Floating point registers and status, as saved
  *	and restored by FP save/restore instructions.
@@ -53,11 +51,11 @@ struct i386_fp_regs {
 					/* space for 8 80-bit FP registers */
 };
 
-#define XSAVE_XCOMP_BV_COMPACT (((uint64_t)1) << 63)
+#define XSAVE_XCOMP_BV_COMPACT (((unsigned long long)1) << 63)
 struct i386_xfp_xstate_header {
-	uint64_t	xfp_features;
-	uint64_t	xcomp_bv;
-	uint64_t	reserved[6];
+	unsigned long long	xfp_features;
+	unsigned long long	xcomp_bv;
+	unsigned long long	reserved[6];
 } __attribute__((packed, aligned(64)));
 _Static_assert(sizeof(struct i386_xfp_xstate_header) == 8*8);
